Resource Integration + Basic Multiplayer
Added a simple script using photon to allow stable connection to an
outside client. Tested locally, seems to work, but am having issues with
inconsistent objects being spawned. Also an error in the HUD integration
of this code, as the basic multiplayer does not yet check who owns
received input, and updates every layer simultaneously. This is very
avoidable, it is just 4am and I need some sleep, will implement
tomorrow. For further info regarding multiplayer with Photon, i'm
following this tut:
http://www.paladinstudios.com/2014/05/08/how-to-create-an-online-multiplayer-game-with-photon-unity-networking/
<currently at the end of the 'spawning a player' section>. Also
re-arranged the player prefab's parenting/childing and moved it to the
resources folder, to allow calls over the network.

***BUG*** PLAYER CLIPS THROUGH FLOOR ON INIT, NO IDEA WHY, ITS 4AM.